AS-3092: Fix farms_import_pkg.operation.

+++ farms-liquibase/scripts/01_00_xx/01_00_03/02/ddl/routines/farms_import_pkg.operation.sql
@@ -114,7 +114,7 @@ begin
             'N',
             ac,
             in_program_year_version_id,
-            chr(64 + op_insert_val.operation_number),
+            chr(64 + op_insert_val.operation_number::int),
             1,
             in_user,
             current_timestamp,
